cleanup users code
